+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All important changes to this project will be documented in this file.
 #### Improvements:
 
 * Added fixing property `log-ips` in `server.properties` when secure authorization enabled on versions `1.20.2` and higher
+* Added fixed translation strings for death messages for version `1.20.4`
 
 #### Fixed: