+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 All important changes to this project will be documented in this file.
 
-### [1.4.5 Unreleased](https://github.com/Druzai/Bot_Mc_discord/compare/1.4.4...1.4.5) (2023-xx-xx)
+### [1.4.5 Unreleased](https://github.com/Druzai/Bot_Mc_discord/compare/1.4.4...1.4.5) (2024-xx-xx)
 
 #### Improvements:
 
 * Added fixing property `log-ips` in `server.properties` when secure authorization enabled on versions `1.20.2` and higher
 * Added fixed translation strings for death messages for version `1.20.4`
 * Added parsing hyperlinks from Markdown, e.g. `[link](http://example.com "title")`
+* Added and updated translation strings for death messages for version `1.20.5`
 
 #### Fixed: